+++ service/imagebuilder/api.go
@@ -14835,10 +14835,10 @@ func (s *StartImagePipelineExecutionOutput) SetRequestId(v string) *StartImagePi
 type SystemsManagerAgent struct {
 	_ struct{} `type:"structure"`
 
-	// This property defaults to true. If Image Builder installs the SSM agent on
-	// a build instance, it removes the agent before creating a snapshot for the
-	// AMI. To ensure that the AMI you create includes the SSM agent, set this property
-	// to false.
+	// Controls whether the SSM agent is removed from your final build image, prior
+	// to creating the new AMI. If this is set to true, then the agent is removed
+	// from the final image. If it's set to false, then the agent is left in, so
+	// that it is included in the new AMI. The default value is false.
 	UninstallAfterBuild *bool `locationName:"uninstallAfterBuild" type:"boolean"`
 }