+++ service/securityhub/doc.go
@@ -4,19 +4,20 @@
 // requests to AWS SecurityHub.
 //
 // Security Hub provides you with a comprehensive view of the security state
-// of your AWS environment and resources. It also provides you with the readiness
-// status of your environment based on controls from supported security standards.
-// Security Hub collects security data from AWS accounts, services, and integrated
-// third-party products and helps you analyze security trends in your environment
-// to identify the highest priority security issues. For more information about
-// Security Hub, see the AWS Security Hub User Guide (https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html) .
+// of your Amazon Web Services environment and resources. It also provides you
+// with the readiness status of your environment based on controls from supported
+// security standards. Security Hub collects security data from Amazon Web Services
+// accounts, services, and integrated third-party products and helps you analyze
+// security trends in your environment to identify the highest priority security
+// issues. For more information about Security Hub, see the Security HubUser
+// Guide (https://docs.aws.amazon.com/securityhub/latest/userguide/what-is-securityhub.html) .
 //
 // When you use operations in the Security Hub API, the requests are executed
-// only in the AWS Region that is currently active or in the specific AWS Region
-// that you specify in your request. Any configuration or settings change that
-// results from the operation is applied only to that Region. To make the same
-// change in other Regions, execute the same command for each Region to apply
-// the change to.
+// only in the Amazon Web Services Region that is currently active or in the
+// specific Amazon Web Services Region that you specify in your request. Any
+// configuration or settings change that results from the operation is applied
+// only to that Region. To make the same change in other Regions, execute the
+// same command for each Region to apply the change to.
 //
 // For example, if your Region is set to us-west-2, when you use CreateMembers
 // to add a member account to Security Hub, the association of the member account